like halfway done to making the rudimentary profile screen

+++ src/main/java/io/github/moosyu/helpers/TextShadowHelper.java
@@ -5,7 +5,7 @@ import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 
 public class TextShadowHelper {
-    public static void drawShadowText(GuiGraphics graphics, String text, int x, int y, int color) {
+    public static void drawShadowText(GuiGraphics graphics, String text, float x, float y, int color) {
         final Font font = Minecraft.getInstance().font;
         int textOffset = (font.width(text) / 2) - 1;
         x -= textOffset;
